refactor: PrepTabScreen 분리
